@@ -28,22 +28,24 @@
         "@astrojs/preact": "^3.5.2",
         "@astrojs/sitemap": "^3.1.6",
         "@astrojs/tailwind": "^5.1.0",
-        "@fontsource/figtree": "^5.0.11",
+        "@fontsource/figtree": "^5.0.21",
         "astro": "^4.15.2",
         "astro-icon": "^1.1.1",
-        "preact": "^10.17.1",
-        "sharp": "^0.32.5",
-        "tailwindcss": "^3.3.3"
+        "preact": "^10.23.2",
+        "sharp": "^0.32.6",
+        "tailwindcss": "^3.4.10"
     },
     "devDependencies": {
-        "@typescript-eslint/eslint-plugin": "^6.6.0",
-        "@typescript-eslint/parser": "^6.6.0",
-        "eslint": "^8.48.0",
-        "eslint-config-prettier": "^9.0.0",
-        "eslint-plugin-prettier": "^5.0.0",
-        "prettier": "^3.0.3",
-        "prettier-plugin-astro": "^0.12.0",
-        "prettier-plugin-tailwindcss": "^0.5.4",
-        "typescript": "^5.2.2"
+        "@eslint/eslintrc": "^3.1.0",
+        "@eslint/js": "^9.9.1",
+        "@typescript-eslint/eslint-plugin": "^8.4.0",
+        "@typescript-eslint/parser": "^8.4.0",
+        "eslint": "^9.9.1",
+        "eslint-config-prettier": "^9.1.0",
+        "eslint-plugin-prettier": "^5.2.1",
+        "prettier": "^3.3.3",
+        "prettier-plugin-astro": "^0.14.1",
+        "prettier-plugin-tailwindcss": "^0.6.6",
+        "typescript": "^5.5.4"
     }
 }
